Moves inspect message function to a separate file
